Flutter/Dart Retrieving Data from Remote APIs with DioLibrary
By pulling data from Remote APIs with Dio Library, I listed the incoming data using FutureBuildirer. After verifying the incoming data by querying it with a snapshot, it was printed on the screen in the form of listTile in Listview. The url information among the remote information was transferred to the leading section of the listTile with networkImage.